簡化 GUI 字型結構並優化元件設計
本次調整簡化了 GUI 字型結構，移除 CHeight 欄位，統一僅保留 YSize 與 Baseline，並同步修正相關建構子與初始化。將 GUI__strlen 回傳型別由 int 改為 uint16_t，提升型別一致性。ProgBar、Radio、Widget 等元件部分私有方法加上 const 修飾，強化物件導向設計。優化 ProgBar 與 Radio 的繪製與計算，移除冗餘變數並修正尺寸計算。移除 main.cpp 中 _UpdateProgBarStatus 及相關呼叫，簡化進度條測試。WObj 結構前置宣告由 struct 改為 class，並修正 WM_CALLBACK 型別。其他細部調整包含修正 include 標頭、初始化變數與註解測試函式呼叫，提升一致性與可維護性。

+++ GUI/Core/GUI_Protected.h
@@ -24,7 +24,7 @@ int GUI__WrapGetNumBytesToNextLine(const char *pText, int xSize, WRAPMODE WrapMo
 
 bool GUI__SetText(char **ppText, const char *s);
 bool GUI__strcmp(const char *s0, const char *s1);
-int GUI__strlen(const char *s);
+uint16_t GUI__strlen(const char *s);
 int GUI__strcpy(char *sDest, const char *sSrc);
 
 void GL_DispChar(uint16_t c);
